requirements.txt
httpx
loguru
aiofiles
python-magic
imagehash
pillow
beautifulsoup4
pandas